+++ src/backend/RequestBackend.ts
@@ -1,17 +1,17 @@
 import RequestContext from "../RequestContext";
 import {ApiResponse}  from "../ApiTypes";
 
-export interface RequestOperation<T> {
-  promise: Promise<ApiResponse<T>>;
+export interface RequestOperation<R> {
+  promise: Promise<R>;
   canceler: () => void;
 }
 
-export default interface RequestBackend<R = any, E extends Error = Error> {
-  makeRequest<T>(context: RequestContext): RequestOperation<T>;
+export default interface RequestBackend<R = any> {
+  makeRequest(context: RequestContext): RequestOperation<R>;
 
   convertResponse<T>(response: R): Promise<ApiResponse<T>>;
 
   extractResponseFromError(
     error: Error,
-  ): Promise<ApiResponse | null | undefined>;
+  ): Promise<R | null | undefined>;
 }