+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - allow defaults to be a function on api
 - add content type header in fetch backend to match axios backend
+- remove `flatted`
+- polyfill object.assign for ie
 
 # 0.2.5